requirements.txt
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
sqlalchemy>=2.0
fastapi>=0.110
uvicorn>=0.29
apscheduler>=3.10
streamlit>=1.33
plotly>=5.20
openpyxl>=3.1
jinja2>=3.1
requests>=2.31
python-dotenv>=1.0
